Optimized minify wrapper function
Previously a single instance of the Minifier was reused between
requests, but this update forces a new one to be created each time.
This actually improves performance, especially in cases where there are
a lot of small scripts being processed, as the cleanup is more
efficient and less checks have to be done to make sure the instance was
cleanly reset.