Add gitattributes w/ export-ignore
This will allow for a production ready version of JShink when archived/exported for release(s) and still allowing a development version within the repository.

The `export-ignore` excludes un-needed files and directories when a branch is archived for release and these files are included when the repo is cloned/forked.

This would be very useful because releases are intended for production use and when users are using a package (via composer or zip), they most likely are not interested in downloading the source code as a whole into production, thus saving server disk space and meeting security protocol.

  - e.g. /test folder, or your .travis.yml file, etc. are not included in the release zip

See: [git-scm.com - Exporting-Your-Repository](http://git-scm.com/book/en/v2/Customizing-Git-Git-Attributes#Exporting-Your-Repository)