Fix where "return/RegEX/" breaks minify tinymce.min.js contains the following Javascript:         return/^[^:]+:\/\/\/?[^\/]+\//.test(t)
JSShrink does not call the saveRegex method for this Javascript string as it is not a recognized RegEx starting point.
By adding "n" to the strpos test before the saveRegex call, the return/RegEx/ situation is properly handled. In the
above case, when it is not properly handled, the \//.test(t) part of the RegEx is treated as a comment and is
improperly truncated in the case of tineymce.min.js.